+++ orquesta/tests/unit/specs/test_base_spec_serialize.py
@@ -12,10 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import copy
 import unittest
 
 from orquesta.tests.unit.specs import base as test_specs
+from orquesta.utils import jsonify as json_util
 
 
 class SpecTest(unittest.TestCase):
@@ -154,19 +154,19 @@ class SpecTest(unittest.TestCase):
         spec_obj_json = spec_obj.serialize()
 
         # Test missing catalog information.
-        test_json = copy.deepcopy(spec_obj_json)
+        test_json = json_util.deepcopy(spec_obj_json)
         test_json.pop('catalog')
 
         self.assertRaises(ValueError, test_specs.MockSpec.deserialize, test_json)
 
         # Test missing version information.
-        test_json = copy.deepcopy(spec_obj_json)
+        test_json = json_util.deepcopy(spec_obj_json)
         test_json.pop('version')
 
         self.assertRaises(ValueError, test_specs.MockSpec.deserialize, test_json)
 
         # Test mismatch version information.
-        test_json = copy.deepcopy(spec_obj_json)
+        test_json = json_util.deepcopy(spec_obj_json)
         test_json['version'] = '99.99'
 
         self.assertRaises(ValueError, test_specs.MockSpec.deserialize, test_json)